Refactor input classes to implement new interfaces.
