+++ src/Input/Base/AbstractChoiceList.php
@@ -6,14 +6,13 @@ namespace PHPForge\Html\Input\Base;
 
 use PHPForge\Html\Attribute;
 use PHPForge\Html\Input\Checkbox;
-use PHPForge\Html\Input\ChoiceInterface;
-use PHPForge\Html\Input\Contract\ClassInterface;
+use PHPForge\Html\Input\Contract\ChoiceInterface;
 use PHPForge\Html\Input\Radio;
 use PHPForge\Html\Label;
 use PHPForge\Html\Tag;
 use PHPForge\Widget\Element;
 
-abstract class AbstractChoiceList extends Element implements ChoiceInterface, ClassInterface
+abstract class AbstractChoiceList extends Element implements ChoiceInterface
 {
     use Attribute\Aria\HasAriaDescribedBy;
     use Attribute\Aria\HasAriaLabel;
