+++ src/Input/Base/AbstractInputChoice.php
@@ -6,8 +6,7 @@ namespace PHPForge\Html\Input\Base;
 
 use PHPForge\Html\Attribute;
 use PHPForge\Html\Input\CheckedValueInterface;
-use PHPForge\Html\Input\ChoiceInterface;
-use PHPForge\Html\Input\Contract\ClassInterface;
+use PHPForge\Html\Input\Contract\ChoiceInterface;
 use PHPForge\Html\Label;
 use PHPForge\Html\Tag;
 use PHPForge\Widget\Element;
@@ -15,7 +14,7 @@ use PHPForge\Widget\Element;
 /**
  * Provides a foundation for creating HTML elements with various attributes and content.
  */
-abstract class AbstractInputChoice extends Element implements CheckedValueInterface, ChoiceInterface, ClassInterface
+abstract class AbstractInputChoice extends Element implements CheckedValueInterface, ChoiceInterface
 {
     use Attribute\Aria\HasAriaDescribedBy;
     use Attribute\Aria\HasAriaLabel;
