--- src/Input/Contract/ClassInterface.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PHPForge\Html\Input\Contract;
-
-/**
- * Provide methods for handling HTML input class attribute.
- */
-interface ClassInterface
-{
-    /**
-     * Set the `CSS` `HTML` class attribute.
-     *
-     * @param string $value The `CSS` attribute of the widget.
-     * @param bool $override If `true` the value will be overridden.
-     *
-     * @return static A new instance of the current class with the specified class value.
-     *
-     * @link https://html.spec.whatwg.org/#classes
-     */
-    public function class(string $value, bool $override = false): static;
-}
